refactor: Rename SubmissionPage to SubmissionsPage
